requirements.txt
Flask
werkzeug
python-dotenv
gunicorn
uwsgi==2.0.30
mysql-connector-python